creates modular files
