+++ index.html
@@ -226,7 +226,7 @@
       </div>
     </section>
   </template>
-  <script src="/js/main.js"></script>
+  <script type="module" src="/js/main.js"></script>
   <script src="/js/functions.js"></script>
 </body>
 </html>
