+++ js/util.js
@@ -1,0 +1,4 @@
+const getRandomNumberFromRange = (min,max) =>
+  Math.floor(Math.random() * (max - min)) + min;
+
+export {getRandomNumberFromRange};
